requirements.txt
arrow==0.7.0
beautifulsoup4==4.4.1
cffi==1.5.0
characteristic==14.3.0
cryptography==1.2.1
cssselect==0.9.1
Django==1.9.1
enum34==1.1.2
idna==2.0
ipaddress==1.0.16
iso8601==0.1.11
leancloud-sdk==1.3.12
lxml==3.5.0
Pillow==3.1.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pyOpenSSL==0.15.1
python-dateutil==2.4.2
qiniu==7.0.6
queuelib==1.4.2
requests==2.9.1
Scrapy==1.0.4
scrapy-qiniu==0.1.2
scrapyd==1.1.0
scrapyd-client==1.0.1
service-identity==14.0.0
six==1.10.0
Twisted==15.5.0
w3lib==1.13.0
Werkzeug==0.11.3
wheel==0.24.0
zope.interface==4.1.3
user-agent==0.1.1
